Refactor cosmos initialization so that it can be used in tests
